Extract webhook logs handlers to separate file
